dev terraform bucket

+++ scripts/deleteBucket.sh
@@ -1,2 +1,9 @@
 #!/bin/bash
-aws s3 rm  s3://f5-public-cloud-terraform-${AWS_DEFAULT_REGION}/${TF_VAR_emailidsan}.terraform.tfstate
+if [ $F5_ENV == "development" ]
+  then
+  aws s3 rm  s3://f5-public-cloud-development-terraform-${AWS_DEFAULT_REGION}/${TF_VAR_emailidsan}.terraform.tfstate
+  else
+  aws s3 rm  s3://f5-public-cloud-terraform-${AWS_DEFAULT_REGION}/${TF_VAR_emailidsan}.terraform.tfstate
+fi
+
+
